feat: add configuration of the project
- add environments folder
