+++ .prettierrc.js
@@ -1,6 +1,9 @@
 module.exports = {
-  bracketSpacing: false,
+  bracketSpacing: true,
   jsxBracketSameLine: true,
   singleQuote: true,
   trailingComma: 'all',
+  tabWidth: 2,
+  semi: true,
+  printWidth: 140,
 };
