+++ environments/.env.development
@@ -1,0 +1,2 @@
+# App Variables
+API_BASE_URL=http://cas-entertainment.dev/api/
