+++ jest.config.json
@@ -1,0 +1,22 @@
+{
+  "preset": "react-native",
+  "roots": ["<rootDir>/src/"],
+  "testMatch": ["**/+(*.)+(spec).+(ts)"],
+  "collectCoverage": true,
+  "collectCoverageFrom": ["**/*ts", "!**/*-mock.ts"],
+  "coverageDirectory": "coverage",
+  "setupFiles": ["./jest/setup.js"],
+  "moduleFileExtensions": [
+    "ts",
+    "tsx",
+    "js",
+    "jsx",
+    "json",
+    "node"
+  ],
+  "globals": {
+    "ts-jest": {
+      "tsConfig": "./tsconfig.spec.json"
+    }
+  }
+}
